telegram_weather_api/pyproject.toml
[tool.poetry]
name = "telegram_weather_api"
version = "0.1.0"
description = ""
authors = ["Max <brunomax23@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
django-rest-framework = "^0.1.0"
django-environ = "^0.11.2"
psycopg = "^3.1.18"
django-redis = "^5.4.0"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
black = "^24.1.1"

[tool.black]
line-length = 110

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
